observes the "time-pos" property instead of using timers
So this is what I've finally got working! Instead of using calculated timers which we have to make a bunch of catches to make sure it succeeded, this just observes the "time-pos" property and restores when our calculated time is reached. I implemented your latest "desync" commit and it improves the performance here as well!