Restore some required logic and refactor
The state machine is strictly required for the script to work properly. When seeking back in speedup state and a sub becomes visible, the code should normally restore normalspeed, but it doesn't because it doesn't know that it is in speedup mode. That's why it is necessary to know in which state we currently are. Also remove some global vars and make variables local where possible.