delete_file: fix for files not deleted on end of playlist
Currently, the selected files are deleted only when mpv is
exited via the quit command. If, however, the quit reason is
eof of the last file in the playlist, delete_file would not
delete any files.

To fix this use the correct event, i.e. shutdown, instead of
end-file.

See:

https://github.com/mpv-player/mpv/blob/master/DOCS/man/lua.rst#list-of-events