rework skip back code and conditions
In order to not always needlessly skip all the way back if we skipped anywhere to where a subtitle is in the process of being played back we introduce a margin.
This margin defines if we are far enough of the end time of a subtitle to not skip back. So if we skipped near the end of a subtitle we skip back to the previous skip position. But if we are likely at the start of a subtitle we don't skip back.

This reduces frequent annoying skip backs.

TODO: make this maybe configurable or adjust the margin