Speed transition fixes/improvements (#57)
* skipval() now returns if skipping is possible

adjust code and remove unneeded checks

* adjust skip mode in state 2 to allow skipping within the leadin window

also rename nextsub to more appropriate nextsub_start

* fix last_skip_position: int state 3 position was stored before the delay

but skipping is done after the delay. Fix this so when skipping back now jumps to correct position

* rework skip back code and conditions

In order to not always needlessly skip all the way back if we skipped anywhere to where a subtitle is in the process of being played back we introduce a margin.
This margin defines if we are far enough of the end time of a subtitle to not skip back. So if we skipped near the end of a subtitle we skip back to the previous skip position. But if we are likely at the start of a subtitle we don't skip back.

This reduces frequent annoying skip backs.

TODO: make this maybe configurable or adjust the margin

* in skip mode if we are within the leadin window dont' skip back

* fix check_should_speedup getting wrong time for subend in state 3

we were giving the actual position in time for check_should_speedup() but it expects the subtitle end position.
So pass it the sub-end property (this is often still nil however) so check_should_speedup() can properly work.

* add/remove/change some debug info + cleanup

* set last_speedup_zone_begin to 0 on reset_on_file_load()

* sleep() after sub-step to get fresh data

after sub-stepping mpv would often return stale values which impacted the working of the script negatively.
Waiting sometime or just printing the return value would get us the correct value.
Therefore we sleep() after stepping.

* check_should_speedup(): add handling for special case of lookahead = 0

if lookahead = 0 the script would always speedup regardless of cfg.leadin.

* check_should_speedup(): reduce the amount of state switching when in skipmode if conditions for skipping are not met.

* make the rest of the keybindings also repeatable

* uncomment wait_finish_seeking()